pyproject.toml
[project]
name = "automated-pipeline"
version = "0.1.0"
description = "Automated software delivery pipeline using AI-driven workflow orchestration"
authors = [
    {name = "Ben", email = "ben@example.com"},
]
dependencies = [
    "pydantic>=2.0.0",
    "structlog>=23.1.0",
    "pyyaml>=6.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
pipeline = "pipeline.cli:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]
